+++ fixtures/callbacks/tests/test_generated_bindings.rs
@@ -3,6 +3,6 @@ uniffi_macros::build_foreign_language_testcases!(
     [
         "tests/bindings/test_callbacks.kts",
         "tests/bindings/test_callbacks.swift",
-        //"tests/bindings/test_callbacks.py", // see https://github.com/mozilla/uniffi-rs/pull/1068
+        "tests/bindings/test_callbacks.py",
     ]
 );