+++ uniffi_bindgen/src/bindings/kotlin/templates/CallbackInterfaceTemplate.kt
@@ -20,18 +20,33 @@ public interface {{ type_name }} {
 // The ForeignCallback that is passed to Rust.
 internal class {{ foreign_callback }} : ForeignCallback {
     @Suppress("TooGenericExceptionCaught")
-    override fun invoke(handle: Handle, method: Int, args: RustBuffer.ByValue): RustBuffer.ByValue {
+    override fun invoke(handle: Handle, method: Int, args: RustBuffer.ByValue, outBuf: RustBufferByReference): Int {
         val cb = {{ ffi_converter }}.lift(handle) ?: throw InternalException("No callback in handlemap; this is a Uniffi bug")
         return when (method) {
-            IDX_CALLBACK_FREE -> {{ ffi_converter }}.drop(handle)
+            IDX_CALLBACK_FREE -> {
+                {{ ffi_converter }}.drop(handle)
+                // No return value.
+                // See docs of ForeignCallback in `uniffi/src/ffi/foreigncallbacks.rs`
+                0
+            }
             {% for meth in cbi.methods() -%}
             {% let method_name = format!("invoke_{}", meth.name())|fn_name -%}
-            {{ loop.index }} -> this.{{ method_name }}(cb, args)
+            {{ loop.index }} -> {
+                val buffer = this.{{ method_name }}(cb, args)
+                outBuf.setValue(buffer)
+                // Value written to out buffer.
+                // See docs of ForeignCallback in `uniffi/src/ffi/foreigncallbacks.rs`
+                1
+            }
             {% endfor %}
             // This should never happen, because an out of bounds method index won't
             // ever be used. Once we can catch errors, we should return an InternalException.
             // https://github.com/mozilla/uniffi-rs/issues/351
-            else -> RustBuffer.ByValue()
+            else -> {
+                // An unexpected error happened.
+                // See docs of ForeignCallback in `uniffi/src/ffi/foreigncallbacks.rs`
+                -1
+            }
         }
     }